requirements.txt
streamlit==1.32.0
torch==2.1.2
torchaudio==2.1.2
transformers==4.36.2
accelerate==0.25.0
soundfile==0.12.1
huggingface_hub==0.20.3
numpy==1.26.4
python-dotenv==1.0.1
